"fixed" reference loops and added example
Reference loops on serialized-by-value objects now dont cause a stack overflow (that's good). However, they are *ignored*, not managed in any fancy way (that's bad), so you still probably don't want to have them.

Also finally got around to creating an example for how to serialize anything by reference. It's a bit of work, and not terribly easy if you are a beginner.

But hey, it works.